refactor: `List<Line>`을 감싸는 Lines 삭제함
