+++ src/main/java/ladder/domain/Ladder.java
@@ -1,5 +1,7 @@
 package ladder.domain;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -8,30 +10,50 @@ public class Ladder {
 
     private static final String EMPTY_LADDER_EXCEPTION_MESSAGE = "사다리는 Line이 1개 이상 있어야 합니다";
 
-    private final Lines lines;
+    private final List<Line> lines;
 
     private Ladder(List<Line> lines) {
-        this(Lines.from(lines));
-    }
-
-    private Ladder(Lines lines) {
         this.lines = lines;
     }
+
 
     public static Ladder from(List<Line> lines) {
         if (lines.isEmpty()) {
             throw new IllegalArgumentException(EMPTY_LADDER_EXCEPTION_MESSAGE);
         }
-        return new Ladder(lines);
+        List<Line> unmodifiableLines = Collections.unmodifiableList(lines);
+        return new Ladder(unmodifiableLines);
     }
 
     public LadderResult result() {
-        Map<Integer, Integer> resultIndices = lines.resultIndices();
+        Map<Integer, Integer> resultIndices = resultIndices();
         return new LadderResult(resultIndices);
     }
 
+    private Map<Integer, Integer> resultIndices() {
+        Map<Integer, Integer> resultIndices = new HashMap<>();
+        for (int i = 0; i < lines.get(0).size(); i++) {
+            int resultIndex = resultIndexOf(i);
+            resultIndices.put(i, resultIndex);
+        }
+        return resultIndices;
+    }
+
+    private int resultIndexOf(int index) {
+        int lineIndex = 0;
+        while (lineIndex < lines.size()) {
+            System.out.printf("%d %d\n", lineIndex, index);
+            Line currentLine = lines.get(lineIndex);
+            Link currentLink = currentLine.linkOf(index);
+
+            lineIndex++;
+            index = currentLink.nextIndex(index);
+        }
+        return index;
+    }
+
     public List<Line> lines() {
-        return lines.value();
+        return lines;
     }
 
     @Override
